pyproject.toml
[project]
name = "advanced-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "firecrawl-py>=4.18.1",
    "langchain>=1.2.10",
    "langchain-core>=1.2.17",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
